Add multipath route support for interfaces using method dhcp.
Added policy rules table support, and this is checked firstly.  And if
supported, the interface table-name is obtained.

When multiple default routes are desired for multiple interfaces, they can
be added to the routing table as a 'nexthop'.

A "route weight" value is applied to each 'nexthop', based on the ifindex,
if not otherwise defined.

If the kernel does not include IP_ROUTE_MULTIPATH, normal (exclusive)
default routing behavior applies.

Note: to also address Bug 5578...
For 'renew' action, the ip-address is modified in-place and routes are
replaced. This should help avoid routing delay on-change.

Bugs 4973, 5578.