Simplify client check on renew request.
Moved client check to follow-up on a renew request, instead of before it.
The check is just a signal 0, to see if the client is still responsive.

This deals with the udhcpc client abruptly quitting (bug) upon sending a
renew request after recreation of the interface phy.  This scenario will
occurs after a 'wireless restart'.

ifrc -- v20140716

Bug 5080.