Wait for wireless setup to complete, before applying L2 rules.
Add dynamic wait for wireless setup to complete.
Then check if the supplicant's socket exists, to determine the need
to apply the --arpnat rules. This allows for automatically applying
the rules only when using the supplicant, which needs them.

Bug 6185.
Fix default ssid handling for hostapd.

In wireless.sh, for preparing to launch hostapd...
Detect if hostapd.conf has an unset ssid like: 'wb??n_id-not-set'

If match, then re-write this ssid with format: 'wb??n_<radio-mac-vei>'

Example written default is:
ssid=wb45n_e01042

The sed command will modify the ssid line, editing the file in-place.
And then the hostapd.conf is fsync'd.

Bug 6199, 6206.