Laird: wb40: Use a live-kernel tree for the development target
This switches the wb40n_devel target over to utilizing a live-tree for the
kernel, the same way we do development on the wb45n_devel target.

One caveat: be sure to switch the kernel tree in
package/lrd-closed-source/externals/kernel to the necessary branch for
the target you're trying to build.

Bug: 6924