Fix ask/reply logic for proceeding with the overwrite of a same-image.
The expected reply for confirming to overwrite is '[y/N]', meaning "no" is
to be assumed for a non-reply. It should only proceed with an overwrite if
the keypress is 'y', meaning "yes". However, if some other key is captured,
then the assumed "no" is actually avoided ...which is incorrect.

Behavior is fixed by inverting the test, requiring the reply to be "yes".

Bug 7381