Add background re-exec mechanism to avoid blocking call via netlink daemon.
The netlink daemon uses a spawn_and_wait() to call the run_script, and
this can cause both the daemon and ifrc to be "blind" to a changed status.

The re-exec mechanism works as follows:

When ifrc "sees" an ampersand character as the last arg, it will re-exec
itself to run in the background. This same character is added to the ifnl
invocation, such that calls via the netlink daemon run_script will result
in a non-blocking call to ifrc. (Critical operations are already protected
with temporary lock-files, so conflicts are thus avoided.)

Manual invocations by the user or other scripted calls (w/o the ampersand)
will continue to work as before.

  ifrc -- v20140910

Bug 7625