laird: Disable vulnerable ssh client roaming code
Turns off the possiblity of ssh roaming for the client to block
CVE-2016-0777 and CVE-2016-0778.

This is only applied to GA4 branches as a low-risk fix. GA5 and
master branches will upgrade ssh. GA3 and older branches won't
include this and customers will be given instructions on how to
patch their builds.

Bug: 8883