laird: wbXXn: don't restrict eth0 phy
If the kernel has trouble identifying the eth0 phy as one of our supported
drivers, it falls back to the Generic phy driver. Many typical phy chips
can be successfully used with this driver. So it makes no sense to force
our network scripts to abort if we see this driver loaded.

It is true that our scripts might try to "up" a non-existent phy,
but that is OK and other situations can detect and debug that error just as
readily.

This change to our scripts removes the checking and abort code when a
Generic phy driver is found to allow other phy chips to be utilized.

Bug: 9291