Add kernel size check
Decision has been made to enforce a 2 spare block requirement on
the WB45 kernel at build time.  Need to update the build scripts
to check the size of the kernel image and fail the build if it
exceeds the 18 block limit (2359296 bytes). The checking is added
in post_image.sh.

bug: 15337