Reserve additional blocks in UBI partition
Previous implementation reserved 1 percent of blocks in UBI partition for bad blocks. This resulted in only 2 blocks available as spares.
Modern kernels reserve approximately 2 percent of all blocks in the chip for each UBI device, as blocks could go bad in clusters.
Reserve 10 percent of blocks in UBI partition which results in a total of 20 blocks.  This is inline with modern kernels, and still leaves slightly more rootfs space than original design due to lower overhead  with larger blocksize.

Bug: 16574