Revert non-adaptive-ww portions of "buildroot: Changed SOM60/WB50 FIPS self-test to GCM"
The reason these commits must be cherry-picked to lrd-7.625.0.x is because
moving Alcon to a 10.135.0.x radio release breaks the build due to changes in
adaptive_ww buildroot repo.  The reason Alcon must move to a 10.135.0.x radio
release is due to dnsmasq not serving DHCP addresses after moving Alcon to
Summit Firewall.

Bug 23505

This reverts commit 9637816ea46a1f0e0dd9e1078d844c30b18d6dcc.